add user login

+++ app/Http/Controllers/Frontend/CartController.php
@@ -11,6 +11,7 @@ use App\Models\AdManager;
 use App\Models\Product;
 use Illuminate\Support\Facades\Validator;
 use RealRashid\SweetAlert\Facades\Alert;
+use Illuminate\Support\Facades\Auth;
 
 class CartController extends Controller
 {
@@ -23,9 +24,9 @@ class CartController extends Controller
     {
         $categories = Category::with('get_child_category')->get();
         $ads = AdManager::all();
-        $count = WishList::select('id')->count();
-        $cart = Cart::latest()->where('user_id',auth()->user()->id)->get();
-        $count1 = Cart::select('id')->count();
+        $cart = Cart::latest()->where('user_id',auth()->user()->id ?? '')->get();
+        $count = WishList::select('id')->where('user_id',auth()->user()->id ?? '')->count();
+        $count1 = Cart::select('id')->where('user_id',auth()->user()->id ?? '')->count();
 
         return view('layouts.frontend.cart.cart_list',[
             'ads'=>$ads,
@@ -54,31 +55,39 @@ class CartController extends Controller
      */
     public function store(Request $request)
     {
-        $product = Product::where('slug',$request->slug)->with('get_product_avatars')->first();
-        $wish = Cart::where('product_id',$product->id)->first();
-        if ($wish) {
-            return response()->json([
-                'errors'=> 'error'
-            ]);
-        }else{
-            
-
-            $data = Cart::create([
-                'product_id'=> $product->id,
-                'user_id'=> auth()->user()->id
-            ]);
-
-            if ($data) {
-                WishList::where('product_id',$request->id)->delete();
-                $count = WishList::select('id')->count();
-                $count1 = Cart::select('id')->count();
-
+        if(Auth::check()){
+            $product = Product::where('slug',$request->slug)->with('get_product_avatars')->first();
+            $wish = Cart::where('product_id',$product->id && 'user_id',auth()->user()->id)
+            ->first();
+            if ($wish) {
                 return response()->json([
-                    'count'=>$count,
-                    'count1'=>$count1
+                    'errors'=> 'error'
                 ]);
+            }else{
+                
+
+                $data = Cart::create([
+                    'product_id'=> $product->id,
+                    'user_id'=> auth()->user()->id
+                ]);
+
+                if ($data) {
+                    WishList::where('product_id',$request->id)->delete();
+                    $count = WishList::select('id')->count();
+                    $count1 = Cart::select('id')->count();
+
+                    return response()->json([
+                        'count'=>$count,
+                        'count1'=>$count1
+                    ]);
+                }
             }
+        }else{
+            return response()->json([
+                'guest'=>'guest'
+            ]);
         }
+        
     }
 
     /**
@@ -124,8 +133,8 @@ class CartController extends Controller
     public function destroy(Request $request)
     {
         Cart::find($request->id)->delete();
-        $cart = Cart::where('user_id',auth()->user()->id)->get();
-        $count1 = Cart::select('id')->count();
+        $cart = Cart::where('user_id',auth()->user()->id ?? '')->get();
+        $count1 = Cart::select('id')->where('user_id',auth()->user()->id ?? '')->count();
         toast('Product remove successfully','success')->padding('10px')->width('270px')->timerProgressBar()->hideCloseButton();
 
         return response()->json([
